renaming and replacing (refactored) "processFileName" with "cleanFileName" which does all the same things but also handles included full/partial paths and truncating filenames that are too long
