--- packages/common/src/utils/processFileName.ts
@@ -1,47 +1,0 @@
-/*
-CPAL-1.0 License
-
-The contents of this file are subject to the Common Public Attribution License
-Version 1.0. (the "License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-https://github.com/ir-engine/ir-engine/blob/dev/LICENSE.
-The License is based on the Mozilla Public License Version 1.1, but Sections 14
-and 15 have been added to cover use of software over a computer network and 
-provide for limited attribution for the Original Developer. In addition, 
-Exhibit A has been modified to be consistent with Exhibit B.
-
-Software distributed under the License is distributed on an "AS IS" basis,
-WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
-specific language governing rights and limitations under the License.
-
-The Original Code is Infinite Reality Engine.
-
-The Original Developer is the Initial Developer. The Initial Developer of the
-Original Code is the Infinite Reality Engine team.
-
-All portions of the code written by the Infinite Reality Engine team are Copyright © 2021-2023 
-Infinite Reality Engine. All Rights Reserved.
-*/
-
-/**
- * This method perform different actions on filename to ensure it meets all requirements.
- * @param fileName File name to be processed
- */
-export const processFileName = (fileName: string): string => {
-  let name = fileName
-
-  try {
-    // Change file extension to lowercase
-    const nameSplit = name.split('.')
-    if (nameSplit.length === 1) return name
-    const extension = nameSplit.pop()
-    if (extension) {
-      nameSplit.push(extension.toLowerCase())
-      name = nameSplit.join('.')
-    }
-  } catch (e) {
-    return name
-  }
-
-  return name
-}
